Feat/permission (#94)
* feat: permission guard and on route back

- Creation d'un Guard pour empêcher l'accès aux routes
-Création d'un service Permission pour permettre d'obtenir les
permissions liés aux objets (GNM_SITES etc) du module Monitorings et
ainsi "bloquer" l'action sur les composants
- Ajout des "check_cruved_scope" coté backend pour empêcher l'éxécution
  d'action CRUD directement depuis les routes API

WIP: continuer à implémenter coté frontend l'affichage conditionnel en
fonction des permissions de l'utilisateurs (affichage ou non des tabs
sites, groupes de site, grisage et désactivation des icones liées aux
actions CRUD, désactivation conditionnelle des boutons liés aux actions
CRUD


* feat: add ux (permission) icon and button

- UX permission to not display sites or sites_groups if not allowed
- UX permission to disable button/icon and add tooltip to warn the user
  that he doesn't get rights to do the action

WIP: add UX to button delete in Form component


* feat: add scope_filter to monitorings obj permiss

Set true in revision alembic to allow admin to choose 'mon organisme' ,
'mes données'


* feat: add variable config in json

Rendre la possibilité d'extraire les variables de configs depuis le
front et de les remplacer par les valeurs de la config.

FIX: bug de redirection à la création d'un site à partir de l'onglet
"sites"


* feat: permission with paginate

Create paginate_scope in order to send object with permission for each
object

Create function model and function query to filter on object and check
scope (based on TDataset model)

Integration of permission in data-table


* feat: add filter readable query to geom site

Add filter_by_readable for site geom


* feat(back): add object_code to BibTypeSites

Specify object_code to BibTypeSite in order to make work permission on
this object


* feat(back): permission on different object

Add permission for:

- GET sites_group geom
- GET sites_group
- GET submodules list when we want to create visit through site entrance

Create PermissonModel to use function created in GN Core

Create function with parameter list object, module_code(opt), object_code
(opt) in order to return CRUVED for each object of list object (need to
test to see if works with any object : Modules --> ok, sites,
sites_group, visits,observations to check


* feat: change the way to get cruved_object

fix problem from frontend to GET cruved_object with object of boolean
(no more object of integer)


* feat: add permission logic (entrance site)

WIP : try to make working VisitQuery readable for organism


* feat: query visit filtering scope

Fix problem function filter_by_scope by using Models


* style(backend): apply black


* feat(back): add observation query and fix

Fix: has/any on filter_by_scope function

Add: observation query and id_digitiser to model


* feat(back): R,U,D for specific object scope

Prevent user to R,U,D for specific object according to his scope


* fix: error on route

default id not needed


* feat: permission entrance through protocol

Front
- Remove unused 'userCruved' (front)
- Set permission for object and child with 'cruved' object get from
  backend (disable btn and icon everywhere cruved is checked)

Back:
- add boolean cruved object for each objectType inside serializer
- add ObservtionQuery to model TMOnitoringOBservation and hybrid_property and has_instance_permission method
- FIX:
   * name "visite" to "visit"
   * change get_id by get_id_name for GenericModel
   * problem inside get_object_with_permission_boolean

WIP:
- Entrance protocol :
  * filter_by_readable "GET" on list object
  * Prevent access to edit page
- Entrance site:
  * UX according to permission on button "Supprimer" into Form Component


* feat: filter on only readable object

Entrance through protocol : filter_by_scope (Readable object)


* feat: prevent user to edit object if not allowed

Add boolean canUpdate in form component in case object is readable but
not editable and user change param edit=true in ulr directly


* fix: add module_code in get_readable_list_object


* feat: add check permission to route (entrance site)

BACK:
- Add to all route  decorator permission and check
  has_intance_permission (entrance site)
- Add 'cruved' to all object return for get_by_id (site and sites_group)
FIX (BACK):
- error when no child (check len childs_model)
- method get_object_with_permission_boolean condition for module need to
  change the call method

FRONT:
- change canUpdate, canDelete for form and property component in order
  to check if specific object has permission
FIX:
- initForm need to wait for 'config' loaded for this.obj
- same thing for form.service


* feat: add guard for other frontend route

Add guard for specific route
Check canRead with canCreate

+++ backend/gn_module_monitoring/routes/sites_groups.py
@@ -1,13 +1,16 @@
-from flask import jsonify, request
+from flask import jsonify, request,g
 from geonature.utils.env import db
 from marshmallow import ValidationError
 from sqlalchemy import func
 from werkzeug.datastructures import MultiDict
-
+from werkzeug.exceptions import Forbidden
+from geonature.core.gn_permissions import decorators as permissions
 from gn_module_monitoring.blueprint import blueprint
 from gn_module_monitoring.config.repositories import get_config
 from gn_module_monitoring.modules.repositories import get_module
 from gn_module_monitoring.monitoring.definitions import monitoring_definitions
+from gn_module_monitoring import MODULE_CODE
+from geonature.core.gn_permissions.decorators import check_cruved_scope
 from gn_module_monitoring.monitoring.models import TMonitoringSites, TMonitoringSitesGroups
 from gn_module_monitoring.monitoring.schemas import MonitoringSitesGroupsSchema
 from gn_module_monitoring.utils.errors.errorHandler import InvalidUsage
@@ -17,7 +20,9 @@ from gn_module_monitoring.utils.routes import (
     get_limit_page,
     get_sort,
     paginate,
+    paginate_scope,
     sort,
+    get_objet_with_permission_boolean
 )
 from gn_module_monitoring.routes.monitoring import (
     create_or_update_object_api_sites_sites_group,
@@ -32,8 +37,10 @@ def get_config_sites_groups(id=None, module_code="generic", object_type="sites_g
     return obj["properties"]
 
 
-@blueprint.route("/sites_groups", methods=["GET"])
-def get_sites_groups():
+@blueprint.route("/sites_groups", methods=["GET"], defaults={"object_type": "sites_group"})
+@check_cruved_scope("R", module_code=MODULE_CODE, object_code="GNM_GRP_SITES")
+def get_sites_groups(object_type: str):
+    object_code = "GNM_GRP_SITES"
     params = MultiDict(request.args)
     limit, page = get_limit_page(params=params)
     sort_label, sort_dir = get_sort(
@@ -42,25 +49,47 @@ def get_sites_groups():
     query = filter_params(query=TMonitoringSitesGroups.query, params=params)
 
     query = sort(query=query, sort=sort_label, sort_dir=sort_dir)
-    return paginate(
-        query=query,
-        schema=MonitoringSitesGroupsSchema,
-        limit=limit,
-        page=page,
+
+    query_allowed = query.filter_by_readable(object_code=object_code)
+    return paginate_scope(
+        query=query_allowed, schema=MonitoringSitesGroupsSchema, limit=limit, page=page
     )
+    # return paginate(
+    #     query=query,
+    #     schema=MonitoringSitesGroupsSchema,
+    #     limit=limit,
+    #     page=page,
+    # )
 
 
-@blueprint.route("/sites_groups/<int:id_sites_group>", methods=["GET"])
-def get_sites_group_by_id(id_sites_group: int):
+@blueprint.route(
+    "/sites_groups/<int:id_sites_group>", methods=["GET"], defaults={"object_type": "sites_group"}
+)
+@check_cruved_scope("R", module_code=MODULE_CODE, object_code="GNM_GRP_SITES")
+@permissions.check_cruved_scope(
+    "R", get_scope=True, module_code=MODULE_CODE, object_code="GNM_GRP_SITES"
+)
+def get_sites_group_by_id(scope,id_sites_group: int, object_type: str):
+    sites_group = TMonitoringSitesGroups.query.get_or_404(id_sites_group)
+    if not sites_group.has_instance_permission(scope=scope):
+        raise Forbidden(f"User {g.current_user} cannot read site group {sites_group.id_sites_group}")
     schema = MonitoringSitesGroupsSchema()
     result = TMonitoringSitesGroups.query.get_or_404(id_sites_group)
-    return jsonify(schema.dump(result))
+    response = schema.dump(result)
+    response['cruved']=get_objet_with_permission_boolean([result], object_code="GNM_GRP_SITES")[0]['cruved']
+    return jsonify(response)
 
 
-@blueprint.route("/sites_groups/geometries", methods=["GET"])
-def get_sites_group_geometries():
+@blueprint.route(
+    "/sites_groups/geometries", methods=["GET"], defaults={"object_type": "sites_group"}
+)
+@check_cruved_scope("R", module_code=MODULE_CODE, object_code="GNM_GRP_SITES")
+def get_sites_group_geometries(object_type: str):
+    object_code = "GNM_GRP_SITES"
+    query = TMonitoringSitesGroups.query
+    query_allowed = query.filter_by_readable(object_code=object_code)
     subquery = (
-        db.session.query(
+        query_allowed.with_entities(
             TMonitoringSitesGroups.id_sites_group,
             TMonitoringSitesGroups.sites_group_name,
             func.st_convexHull(func.st_collect(TMonitoringSites.geom)),
@@ -78,18 +107,34 @@ def get_sites_group_geometries():
     return jsonify(result)
 
 
-@blueprint.route("/sites_groups/<int:_id>", methods=["PATCH"])
-def patch(_id):
+@blueprint.route(
+    "/sites_groups/<int:_id>", methods=["PATCH"], defaults={"object_type": "sites_group"}
+)
+@permissions.check_cruved_scope(
+    "U", get_scope=True, module_code=MODULE_CODE, object_code="GNM_GRP_SITES"
+)
+def patch(scope,_id: int, object_type: str):
     # ###############################""
     # FROM route/monitorings
+    sites_group = TMonitoringSitesGroups.query.get_or_404(_id)
+    if not sites_group.has_instance_permission(scope=scope):
+        raise Forbidden(f"User {g.current_user} cannot update site group {sites_group.id_sites_group}")
+
     module_code = "generic"
-    object_type = "sites_group"
     get_config(module_code, force=True)
     return create_or_update_object_api_sites_sites_group(module_code, object_type, _id), 201
 
 
-@blueprint.route("/sites_groups/<int:_id>", methods=["DELETE"])
-def delete(_id):
+@blueprint.route(
+    "/sites_groups/<int:_id>", methods=["DELETE"], defaults={"object_type": "sites_group"}
+)
+@permissions.check_cruved_scope(
+    "D", get_scope=True, module_code=MODULE_CODE, object_code="GNM_GRP_SITES"
+)
+def delete(scope,_id: int, object_type: str):
+    sites_group = TMonitoringSitesGroups.query.get_or_404(_id)
+    if not sites_group.has_instance_permission(scope=scope):
+        raise Forbidden(f"User {g.current_user} cannot delete site group {sites_group.id_sites_group}")
     item_schema = MonitoringSitesGroupsSchema()
     item = TMonitoringSitesGroups.find_by_id(_id)
     TMonitoringSitesGroups.query.filter_by(id_g=_id).delete()
@@ -97,10 +142,10 @@ def delete(_id):
     return item_schema.dump(item), 201
 
 
-@blueprint.route("/sites_groups", methods=["POST"])
-def post():
+@blueprint.route("/sites_groups", methods=["POST"], defaults={"object_type": "sites_group"})
+@check_cruved_scope("C", module_code=MODULE_CODE, object_code="GNM_GRP_SITES")
+def post(object_type: str):
     module_code = "generic"
-    object_type = "sites_group"
     get_config(module_code, force=True)
     return create_or_update_object_api_sites_sites_group(module_code, object_type), 201
